+++ src/params/types.ts
@@ -28,7 +28,7 @@ import * as logger from "../logger";
  * an Expression<number> as the value of an option that normally accepts numbers.
  */
 export abstract class Expression<T extends string | number | boolean | string[]> {
-  /** Returns the Expression's runtime value, based on the CLI's resolution of params. */
+  /** Returns the expression's runtime value, based on the CLI's resolution of parameters. */
   value(): T {
     if (process.env.FUNCTIONS_CONTROL_API === "true") {
       logger.warn(
@@ -47,11 +47,12 @@ export abstract class Expression<T extends string | number | boolean | string[]>
     throw new Error("Not implemented");
   }
 
-  /** Returns the Expression's representation as a braced CEL expression. */
+  /** Returns the expression's representation as a braced CEL expression. */
   toCEL(): string {
     return `{{ ${this.toString()} }}`;
   }
 
+  /** Returns the expression's representation as JSON. */
   toJSON(): string {
     return this.toString();
   }
@@ -61,8 +62,8 @@ function valueOf<T extends string | number | boolean | string[]>(arg: T | Expres
   return arg instanceof Expression ? arg.runtimeValue() : arg;
 }
 /**
- * Returns how an entity (either an Expression or a literal value) should be represented in CEL.
- * - Expressions delegate to the .toString() method, which is used by the WireManifest
+ * Returns how an entity (either an `Expression` or a literal value) should be represented in CEL.
+ * - Expressions delegate to the `.toString()` method, which is used by the WireManifest
  * - Strings have to be quoted explicitly
  * - Arrays are represented as []-delimited, parsable JSON
  * - Numbers and booleans are not quoted explicitly
@@ -159,7 +160,7 @@ export class CompareExpression<
     return `${this.lhs} ${this.cmp} ${rhsStr}`;
   }
 
-  /** Returns a TernaryExpression which can resolve to one of two values, based on the resolution of this comparison. */
+  /** Returns a `TernaryExpression` which can resolve to one of two values, based on the resolution of this comparison. */
   thenElse<retT extends string | number | boolean | string[]>(
     ifTrue: retT | Expression<retT>,
     ifFalse: retT | Expression<retT>
@@ -171,63 +172,120 @@ export class CompareExpression<
 /** @hidden */
 type ParamValueType = "string" | "list" | "boolean" | "int" | "float" | "secret";
 
+/** Create a select input from a series of values. */
+export function select<T>(options: T[]): SelectInput<T>;
+
+/** Create a select input from a map of labels to vaues. */
+export function select<T>(optionsWithLabels: Record<string, T>): SelectInput<T>;
+
+/** Create a select input from a series of values or a map of labels to values */
+export function select<T>(options: T[] | Record<string, T>): SelectInput<T> {
+  let wireOpts: SelectOptions<T>[];
+  if (Array.isArray(options)) {
+    wireOpts = options.map((opt) => ({ value: opt }));
+  } else {
+    wireOpts = Object.entries(options).map(([label, value]) => ({ label, value }));
+  }
+  return {
+    select: {
+      options: wireOpts,
+    },
+  };
+}
+
+/** Create a multi-select input from a series of values. */
+export function multiSelect(options: string[]): MultiSelectInput;
+
+/** Create a multi-select input from map of labels to values. */
+export function multiSelect(options: Record<string, string>): MultiSelectInput;
+
+/** Create a multi-select input from a series of values or map of labels to values. */
+export function multiSelect(options: string[] | Record<string, string>): MultiSelectInput {
+  let wireOpts: SelectOptions<string>[];
+  if (Array.isArray(options)) {
+    wireOpts = options.map((opt) => ({ value: opt }));
+  } else {
+    wireOpts = Object.entries(options).map(([label, value]) => ({ label, value }));
+  }
+  return {
+    multiSelect: {
+      options: wireOpts,
+    },
+  };
+}
+
 type ParamInput<T> =
-  | { text: TextInput<T> }
-  | { select: SelectInput<T> }
-  | { multiSelect: MultiSelectInput }
-  | { resource: ResourceInput };
+  | TextInput<T>
+  | SelectInput<T>
+  | (T extends string[] ? MultiSelectInput : never)
+  | (T extends string ? ResourceInput : never);
 
 /**
- * Specifies that a Param's value should be determined by prompting the user
- * to type it in interactively at deploy-time. Input that does not match the
+ * Specifies that a parameter's value should be determined by prompting the user
+ * to type it in interactively at deploy time. Input that does not match the
  * provided validationRegex, if present, will be retried.
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export interface TextInput<T = unknown> {
-  example?: string;
-  /**
-   * A regular expression (or an escaped string to compile into a regular
-   * expression) which the prompted text must satisfy; the prompt will retry
-   * until input matching the regex is provided.
-   */
-  validationRegex?: string | RegExp;
-  /**
-   * A custom error message to display when retrying the prompt based on input
-   * failing to conform to the validationRegex,
-   */
-  validationErrorMessage?: string;
+  text: {
+    example?: string;
+    /**
+     * A regular expression (or an escaped string to compile into a regular
+     * expression) which the prompted text must satisfy; the prompt will retry
+     * until input matching the regex is provided.
+     */
+    validationRegex?: string | RegExp;
+    /**
+     * A custom error message to display when retrying the prompt based on input
+     * failing to conform to the validationRegex,
+     */
+    validationErrorMessage?: string;
+  };
 }
 
 /**
- * Specifies that a Param's value should be determined by having the user
+ * Specifies that a parameter's value should be determined by having the user
  * select from a list containing all the project's resources of a certain
  * type. Currently, only type:"storage.googleapis.com/Bucket" is supported.
  */
 export interface ResourceInput {
   resource: {
-    type: string;
+    type: "storage.googleapis.com/Bucket";
   };
 }
 
 /**
- * Specifies that a Param's value should be determined by having the user select
- * from a list of pre-canned options interactively at deploy-time.
+ * Autogenerate a list of buckets in a project that a user can select from.
+ */
+export const BUCKET_PICKER: ResourceInput = {
+  resource: {
+    type: "storage.googleapis.com/Bucket",
+  },
+};
+
+/**
+ * Specifies that a parameter's value should be determined by having the user select
+ * from a list of pre-canned options interactively at deploy time.
  */
 export interface SelectInput<T = unknown> {
-  options: Array<SelectOptions<T>>;
+  select: {
+    options: Array<SelectOptions<T>>;
+  };
 }
 
 /**
- * Specifies that a Param's value should be determined by having the user select
- * a subset from a list of pre-canned options interactively at deploy-time.
- * Will result in errors if used on Params of type other than string[].
+ * Specifies that a parameter's value should be determined by having the user select
+ * a subset from a list of pre-canned options interactively at deploy time.
+ * Will result in errors if used on parameters of type other than `string[]`.
  */
 export interface MultiSelectInput {
-  options: Array<SelectOptions<string>>;
+  multiSelect: {
+    options: Array<SelectOptions<string>>;
+  };
 }
 
 /**
- * One of the options provided to a SelectInput, containing a value and
+ * One of the options provided to a `SelectInput`, containing a value and
  * optionally a human-readable label to display in the selection interface.
  */
 export interface SelectOptions<T = unknown> {
@@ -235,19 +293,19 @@ export interface SelectOptions<T = unknown> {
   value: T;
 }
 
-/** The wire representation of a Param when it's sent to the CLI. A superset of ParamOptions. */
+/** The wire representation of a parameter when it's sent to the CLI. A superset of `ParamOptions`. */
 export type ParamSpec<T extends string | number | boolean | string[]> = {
   /** The name of the parameter which will be stored in .env files. Use UPPERCASE. */
   name: string;
   /** An optional default value to be used while prompting for input. Can be a literal or another parametrized expression. */
   default?: T | Expression<T>;
-  /** An optional human-readable string to be used as a replacement for the Param's name when prompting. */
+  /** An optional human-readable string to be used as a replacement for the parameter's name when prompting. */
   label?: string;
-  /** An optional long-form description of the Param to be displayed while prompting. */
+  /** An optional long-form description of the parameter to be displayed while prompting. */
   description?: string;
   /** @internal */
   type: ParamValueType;
-  /** The way in which the Firebase CLI will prompt for the value of this Param. Defaults to a TextInput. */
+  /** The way in which the Firebase CLI will prompt for the value of this parameter. Defaults to a TextInput. */
   input?: ParamInput<T>;
 };
 
@@ -268,7 +326,7 @@ export type WireParamSpec<T extends string | number | boolean | string[]> = {
   input?: ParamInput<T>;
 };
 
-/** Configuration options which can be used to customize the prompting behavior of a Param. */
+/** Configuration options which can be used to customize the prompting behavior of a parameter. */
 export type ParamOptions<T extends string | number | boolean | string[]> = Omit<
   ParamSpec<T>,
   "name" | "type"
@@ -291,43 +349,43 @@ export abstract class Param<T extends string | number | boolean | string[]> exte
     throw new Error("Not implemented");
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   cmp(cmp: "==" | "!=" | ">" | ">=" | "<" | "<=", rhs: T | Expression<T>) {
     return new CompareExpression<T>(cmp, this, rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   equals(rhs: T | Expression<T>) {
     return this.cmp("==", rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   notEquals(rhs: T | Expression<T>) {
     return this.cmp("!=", rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   greaterThan(rhs: T | Expression<T>) {
     return this.cmp(">", rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   greaterThanOrEqualTo(rhs: T | Expression<T>) {
     return this.cmp(">=", rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   lessThan(rhs: T | Expression<T>) {
     return this.cmp("<", rhs);
   }
 
-  /** Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression. */
+  /** Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression. */
   lessThanOrEqualTo(rhs: T | Expression<T>) {
     return this.cmp("<=", rhs);
   }
 
   /**
-   * Returns a parametrized expression of Boolean type, based on comparing the value of this param to a literal or a different expression.
+   * Returns a parametrized expression of Boolean type, based on comparing the value of this parameter to a literal or a different expression.
    * @deprecated A typo. Use lessThanOrEqualTo instead.
    */
   lessThanorEqualTo(rhs: T | Expression<T>) {
@@ -420,7 +478,7 @@ export class StringParam extends Param<string> {
 /**
  * A CEL expression which represents an internal Firebase variable. This class
  * cannot be instantiated by developers, but we provide several canned instances
- * of it to make available params that will never have to be defined at
+ * of it to make available parameters that will never have to be defined at
  * deployment time, and can always be read from process.env.
  * @internal
  */